CTCP-744: Remove unused object, move DTO to its own file
